dom/nodes/Element-firstElementChild-entity.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="100%" height="100%" version="1.1" viewBox="0 0 400 400"><title>Entity References</title><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><text x="200" y="40" fill="#000" font-size="25" text-anchor="middle">Test of Entity References</text><text id="parentEl" x="200" y="70" fill="#000" font-size="20" text-anchor="middle">The result of this test is &lt;tspan id=&apos;first_element_child&apos; font-weight=&apos;bold&apos;&gt;unknown.&lt;/tspan&gt;</text><h:script><![CDATA[
test(function() {
  var parentEl = document.getElementById("parentEl")
  var fec = parentEl.firstElementChild;
  assert_true(!!fec)
  assert_equals(fec.nodeType, 1)
  assert_equals(fec.getAttribute("id"), "first_element_child")
})
]]></h:script></svg>
